+++ crates/net/eth-wire/src/error.rs
@@ -3,7 +3,7 @@ use std::io;
 
 use reth_primitives::{Chain, ValidationError, H256};
 
-use crate::{capability::SharedCapabilityError, DisconnectReason};
+use crate::{capability::SharedCapabilityError, DisconnectReason, disconnect::UnknownDisconnectReason};
 
 /// Errors when sending/receiving messages
 #[derive(thiserror::Error, Debug)]
@@ -84,6 +84,8 @@ pub enum P2PStreamError {
     SendBufferFull,
     #[error("disconnected")]
     Disconnected(DisconnectReason),
+    #[error("unknown disconnect reason: {0}")]
+    UnknownDisconnectReason(#[from] UnknownDisconnectReason),
 }
 
 // === impl P2PStreamError ===
